SCB-2094 Decouple etcd/util and service pkg
