+++ datasource/etcd/util/versionrule_test.go
@@ -336,78 +336,3 @@ func TestVersionRule(t *testing.T) {
 		assert.Equal(t, false, VersionMatchRule("1.0", "1.6+"))
 	})
 }
-
-func TestNewVersionRegexp(t *testing.T) {
-	log.Info("normal")
-
-	t.Run("latest", func(t *testing.T) {
-		vr := NewVersionRegexp(false)
-		assert.Equal(t, false, vr.MatchString("latest"))
-		vr = NewVersionRegexp(true)
-		assert.Equal(t, true, vr.MatchString("latest"))
-	})
-
-	t.Run("range", func(t *testing.T) {
-		vr := NewVersionRegexp(false)
-		assert.Equal(t, false, vr.MatchString("1.1-2.2"))
-		vr = NewVersionRegexp(true)
-		assert.Equal(t, false, vr.MatchString("-"))
-		assert.Equal(t, false, vr.MatchString("1.1-"))
-		assert.Equal(t, false, vr.MatchString("-1.1"))
-		assert.Equal(t, false, vr.MatchString("1.a-2.b"))
-		assert.Equal(t, false, vr.MatchString("1.-.2"))
-		assert.Equal(t, false, vr.MatchString("60000-1"))
-		assert.Equal(t, true, vr.MatchString("1.1-2.2"))
-		assert.Equal(t, true, vr.MatchString("1.1.1.1-2.2.2.2"))
-		assert.Equal(t, false, vr.MatchString("1.1.1.1.1-2.2.2.2"))
-		assert.Equal(t, false, vr.MatchString("1.1.1.1-2.2.2.2.2"))
-	})
-
-	t.Run("atLess", func(t *testing.T) {
-		vr := NewVersionRegexp(false)
-		assert.Equal(t, false, vr.MatchString("1.0+"))
-		vr = NewVersionRegexp(true)
-		assert.Equal(t, false, vr.MatchString("+"))
-		assert.Equal(t, false, vr.MatchString("+1.0"))
-		assert.Equal(t, false, vr.MatchString("1.a+"))
-		assert.Equal(t, false, vr.MatchString(".1+"))
-		assert.Equal(t, false, vr.MatchString("1.+"))
-		assert.Equal(t, false, vr.MatchString(".+"))
-		assert.Equal(t, false, vr.MatchString("60000+"))
-		assert.Equal(t, true, vr.MatchString("1.0+"))
-		assert.Equal(t, true, vr.MatchString("1.0.0.0+"))
-		assert.Equal(t, false, vr.MatchString("1.0.0.0.0+"))
-	})
-
-	t.Run("explicit", func(t *testing.T) {
-		vr := NewVersionRegexp(false)
-		assert.Equal(t, false, vr.MatchString(""))
-		assert.Equal(t, false, vr.MatchString("a"))
-		assert.Equal(t, false, vr.MatchString("60000"))
-		assert.Equal(t, false, vr.MatchString("."))
-		assert.Equal(t, false, vr.MatchString("1."))
-		assert.Equal(t, false, vr.MatchString(".1"))
-		assert.Equal(t, true, vr.MatchString("1.4"))
-		vr = NewVersionRegexp(true)
-		assert.Equal(t, false, vr.MatchString(""))
-		assert.Equal(t, false, vr.MatchString("a"))
-		assert.Equal(t, false, vr.MatchString("60000"))
-		assert.Equal(t, false, vr.MatchString("."))
-		assert.Equal(t, false, vr.MatchString("1."))
-		assert.Equal(t, false, vr.MatchString(".1"))
-		assert.Equal(t, true, vr.MatchString("1.4"))
-		assert.Equal(t, true, vr.MatchString("1.4.0.0"))
-		assert.Equal(t, false, vr.MatchString("1.4.0.0.0"))
-	})
-
-	log.Info("exception")
-
-	t.Run("MatchString & String", func(t *testing.T) {
-		vr := VersionRegexp{}
-		assert.Equal(t, true, vr.MatchString(""))
-		assert.NotEqual(t, "", vr.String())
-		vr = VersionRegexp{Fuzzy: true}
-		assert.Equal(t, true, vr.MatchString(""))
-		assert.NotEqual(t, "", vr.String())
-	})
-}
